backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
pydantic-settings==2.10.1
bcrypt==4.2.0
email-validator==2.2.0
python-multipart==0.0.20
